Got pyoxidizer kind of working
This is all a big hack, but I wanted to see if it was viable to use
pyoxidizer to package the whisper and numpy together into a single-ish
binary.

The version needed to be a proper semver (0.1.0, not 0.1) in order for
Electron to build at all, and I had to install yarn as a dependency to
get anything to build.

What I did:

Install pyoxidizer: https://github.com/indygreg/PyOxidizer
In the `stagewhisper/app/src/pyoxidizer` directory, run `pyoxidizer
build`

Change the executable file name in `stagewhisper/app/electron/index.ts` to
reference whatever the build comes out on your arch. (I tested in x86_64
on my mac, so it is set to that currently.)

You should now be able to run `npm run dev`, start a `Transcribe`
operation, and it will automatically boot up the Python application.

I haven't gotten it working with the packaged app quite yet (`npm run
dist` mostly works, but the application crashes on startup).

I had to remove the Gooey stuff from the `__main__.py` and move it
around so that the pyxodizer could build it. (PIL doesn't like being
packaged up so neatly.)

Definitely not mergeable as-is, but might give someone else something to
off.